requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
cryptography==3.3.1
decorator==4.4.2
idna==2.10
pycparser==2.20
Pygments==2.7.3
python-telegram-bot==13.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
rich==9.5.1
six==1.15.0
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
